requirements.txt
asgiref==3.8.1
certifi==2024.12.14
distlib==0.3.9
Django==5.1.4
django-debug-toolbar==4.4.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
filelock==3.16.1
packaging==24.2
pipenv==2024.4.0
platformdirs==4.3.6
setuptools==75.6.0
sqlparse==0.5.3
virtualenv==20.28.0
